Changed method used for remote control of the actions. Arrange of code
